Don't copy .gitmodules to $HOME

+++ install.sh
@@ -4,7 +4,7 @@ git submodule init
 git submodule update
 
 for file in .*; do
-  if [ "$file" != "." ] && [ "$file" != ".." ] && [ "$file" != ".git" ]; then
+  if [ "$file" != "." ] && [ "$file" != ".." ] && [ "$file" != ".git" ] && [ "$file" != ".gitmodules" ]; then
     echo "Copying $file"
     rm -rf "$HOME/$file"
     cp -r "$file" "$HOME/$file"
